core: add "nm-errno.h"
This will be our extenion on top of <errno.h>.

We want to use integer error numbers, that can both
contain native errors from <errno.h> and our own defines,
both in the same domain. That is, we will reserve a small
range of integers for our own defines (hopefully not to clash
with errors from <errno.h>).

We can use this at places where GError is too cumbersome to use.

The advantage is, that our error numbers extend <errno.h> and can
be mixed.

This is what "platform/nm-netlink.h" already does. The netlink errors
from there, will be merged with "nm-errno.h".

Also, platform has NMPlatformError, which are a distinct set of errors.
They will be merged as well.

"nm-errno.h" will unify the error handling of platform and netlink,
making it more similar to what we know from systemd, and give room to
extend it for other subsystems.