libnm,cli: add NMSettingWireGuard
For now, only add the core settings.

To support peers and the allowed-ips of the peers, that is more
complicated. It's more complicated because that are nested lists
(allowed-ips) inside a list (peers). To conveniently support that
in keyfile and nmcli is a bit more complicated.

Also, each peer has a secret (the preshared key). Thus we probably need
secret flags for each peer, which is kinda new as well.

That will be added later.