shared,core: move logging enums to header "shared/nm-utils/nm-logging-fwd.h"
In core ("src/"), we use "nm-logging.h" for all logging. This dispatches
to syslog or journal.

If we want to log from a shared component under "shared/", we need to
use a common logging function. Add "nm-utils/nm-logging-fwd.h" for
providing the used logging mechaism.

The shared library will still need to link with "src/nm-logging.c"
or an alternative implementation though. Depending on whether it
is used inside core or somewhere else.