libnm: use "libnm-systemd-shared.a" in "libnm-core.la" (and "libnm.so")
It's not yet used, but it will be. We will need nm_sd_utils_unbase64mem()
to validate WireGuard settings, which contain keys in base64 encoding.

Note that we also need a stub implementation for logging. This just
does nothing for all logging from "libnm-systemd-shared.a". This makes
sense because "libnm.so" as a library should not log directly. Also,
"libnm.so" will only use a small portion of "libnm-systemd-shared.a" which
does not log. Thus this code is anyway unused and dropped by the linker
with "--gc-sections".