libnm: add NMWireGuardEndpoint API
NMWireGuardEndpoint is an immutable structure that contains the endpoint
string. It also does the (naive) parsing of the host and port/service
portions.

NMWireGuardEndpoint allows to have endpoints that are not parsable into
a host and port part. That is useful because our settings need to be
able to hold settings which are invalid. That is for forward compatibility
(server sends a new endpoint format) and for better error handling (have
invalid settings that only later fail the verify() step).