tests: don't use alloca() in tests
The only purpose of using alloca() to avoid the overhead of heap-allocation
and possible save a line in source code for managing/freeing the heap allocation.

For tests we don't care about performance, and (in this case)
the code does not get any shorter.

Avoid alloca() in tests, because alloca() is something to search for
when reviewing code for stack overflows. No need to have such false
positives show up in tests.