libnm: Add async start/stop routines for P2P find operations
These were dropped earlier as new sync API must not be the primary way
of calling new routines in libnm.

In this particular case the DBus calls are simple and unlikely to fail.
Most users should use the normal async API and call the finish routine.
However, if the API user is not interested in the result, then they can
simply set the callback to NULL to ignore it.

[thaller@redhat.com: added options argument to start-find method]